refactor: reinitialized storybook
affects: @atlantis-lab/button, @atlantis-lab/carousel, @atlantis-lab/slideshow,
@atlantis-lab/storybook-config

- Moved config to packages/storybook

ISSUES CLOSED: #151
