+++ packages/storybook/main.js
@@ -1,0 +1,4 @@
+module.exports = {
+  stories: ['../**/src/stories/**/*.story.mdx', '../**/src/stories/**/*.story.@(js|jsx|ts|tsx)'],
+  addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
+}
