+++ packages/storybook/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@atlantis-lab/storybook-config",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "storybook": "start-storybook -p 6006 -c ./",
+    "build-storybook": "build-storybook"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.12.10",
+    "@monstrs/publish-flow": "0.1.3",
+    "@emotion/react": "^11.1.4",
+    "@storybook/addon-actions": "^6.1.15",
+    "@storybook/addon-essentials": "^6.1.15",
+    "@storybook/addon-links": "^6.1.15",
+    "@storybook/react": "^6.1.15",
+    "babel-loader": "^8.2.2",
+    "react": "^16.9.0"
+  }
+}
